moved About & Settings pages to mainScreen
