+++ src/MainScreen/filterHook.tsx
@@ -2,7 +2,7 @@ import { addMonths, subYears } from "date-fns";
 import { useEffect, useState } from "react";
 
 import { convertDateToUnixSeconds } from "../helper";
-import { useSettings } from "../Settings";
+import { useSettings } from "./Settings";
 
 function useFilter() {
